Fixed modifiying command stack when pressing Ctrl+X without selected items

+++ qrgui/editor/editorViewScene.cpp
@@ -697,7 +697,9 @@ void EditorViewScene::deleteSelectedItems()
 		}
 	}
 
-	deleteElements(idsToDelete);
+	if (!idsToDelete.isEmpty()) {
+		deleteElements(idsToDelete);
+	}
 }
 
 void EditorViewScene::deleteElements(IdList &idsToDelete)
